Fixed some typos that caused ped_run's to crash.
There's a point in the ped_run code where a whole
bunch of data fetched from *somewhere* is being
appended to a large number of json elements.
There were a bunch of typos where the same
two pieces of data, qhltemp, and qhlrmstemp, were
being appended to multiple different JSON elements.
Because qhltemp and qhlrmstemp were also JSON nodes,
and each node keeps track of its parent, this didn't
work - the json.c library wouldn't let this happen.
So, I fixed the typo, and now it works. Just in case
there is a need for copying arbitrary JSON nodes,
that function now exists.